Store the recursion level in ctx
Turns out, there were methods for that already. Thanks, the Bouncepaw of the past

+++ mycocontext/interface.go
@@ -51,7 +51,7 @@ func ContextFromStringInput(hyphaName, input string) (Context, CancelFunc) {
 						bytes.NewBufferString(input),
 					),
 					keyRecursionLevel,
-					0,
+					uint(0),
 				),
 				keyWebSiteURL,
 				""),
